# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/application-inventory/review": {
            "get": {
                "description": "List all reviews.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "get"
                ],
                "summary": "List all reviews.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.Review"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a review.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "create"
                ],
                "summary": "Create a review.",
                "parameters": [
                    {
                        "description": "Review data",
                        "name": "review",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Review"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.Review"
                        }
                    }
                }
            }
        },
        "/application-inventory/review/bulk": {
            "post": {
                "description": "Copy a review from one application to others.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "copy"
                ],
                "summary": "Copy a review from one application to others.",
                "parameters": [
                    {
                        "description": "Review copy request data",
                        "name": "copy_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CopyRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
        "/application-inventory/review/{id}": {
            "get": {
                "description": "Get a review by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "get"
                ],
                "summary": "Get a review by ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Review ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.Review"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update a review.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "update"
                ],
                "summary": "Update a review.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Review ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Review data",
                        "name": "review",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Review"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "description": "Delete a review.",
                "tags": [
                    "delete"
                ],
                "summary": "Delete a review.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Review ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
...
        "/controls/stakeholder": {
            "get": {
                "description": "List all stakeholders.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "get"
                ],
                "summary": "List all stakeholders.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.Stakeholder"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a stakeholder.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "create"
                ],
                "summary": "Create a stakeholder.",
                "parameters": [
                    {
                        "description": "Stakeholder data",
                        "name": "stakeholder",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Stakeholder"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.Stakeholder"
                        }
                    }
                }
            }
        },
...
        "/controls/stakeholder/{id}": {
            "get": {
                "description": "Get a stakeholder by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "get"
                ],
                "summary": "Get a stakeholder by ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stakeholder ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Stakeholder"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a stakeholder.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "update"
                ],
                "summary": "Update a stakeholder.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stakeholder ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Stakeholder data",
                        "name": "stakeholder",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Stakeholder"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "description": "Delete a stakeholder.",
                "tags": [
                    "delete"
                ],
                "summary": "Delete a stakeholder.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stakeholder ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.CopyRequest": {
            "type": "object",
            "required": [
                "sourceReview",
                "targetApplications"
            ],
            "properties": {
                "sourceReview": {
                    "type": "integer"
                },
                "targetApplications": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "api.Review": {
            "type": "object",
            "required": [
                "application"
            ],
            "properties": {
                "application": {
                    "type": "object",
                    "required": [
                        "id"
                    ],
                    "properties": {
                        "id": {
                            "type": "integer"
                        }
                    }
                },
                "businessCriticality": {
                    "type": "integer"
                },
                "comments": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "effortEstimate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "proposedAction": {
                    "type": "string"
                },
                "updateUser": {
                    "type": "string"
                },
                "workPriority": {
                    "type": "integer"
                }
            }
        },
        "api.Stakeholder": {
            "type": "object",
            "required": [
                "displayName",
                "email"
            ],
            "properties": {
                "businessServices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BusinessService"
                    }
                },
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "jobFunction": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "integer"
                        },
                        "role": {
                            "type": "string"
                        }
                    }
                },
                "stakeholderGroups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.StakeholderGroup"
                    }
                },
                "updateUser": {
                    "type": "string"
                }
            }
        },
...
        "model.BusinessService": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/model.Stakeholder"
                },
                "ownerID": {
                    "type": "integer"
                },
                "updateUser": {
                    "type": "string"
                }
            }
        },
...
        "model.JobFunction": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "stakeholders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Stakeholder"
                    }
                },
                "updateUser": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "model.Stakeholder": {
            "type": "object",
            "properties": {
                "businessServices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BusinessService"
                    }
                },
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.StakeholderGroup"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "jobFunction": {
                    "$ref": "#/definitions/model.JobFunction"
                },
                "jobFunctionID": {
                    "type": "integer"
                },
                "updateUser": {
                    "type": "string"
                }
            }
        },
        "model.StakeholderGroup": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "stakeholders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Stakeholder"
                    }
                },
                "updateUser": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}